# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MEGA-TIK Finance & Subscribers ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscribers/{id}/renew:
    post:
      operationId: renewSubscription
      tags: [subscribers]
      summary: Renew subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewalRequest"
      responses:
        "200":
          description: Renewal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Renewal"
...
  /renewals:
    get:
      operationId: listRenewals
      tags: [renewals]
      summary: List renewals
      parameters:
        - name: subscriber_id
          in: query
          schema:
            type: integer
        - name: payment_type
          in: query
          schema:
            type: string
            enum: [cash, vodafone_cash, bank_transfer, wallet, deferred]
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of renewals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenewalListResponse"
  /renewals/{id}:
    get:
      operationId: getRenewal
      tags: [renewals]
      summary: Get renewal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Renewal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Renewal"
...
components:
  schemas:
...
    RenewalRequest:
      type: object
      properties:
        planId:
          type: integer
        paymentType:
          type: string
          enum: [cash, vodafone_cash, bank_transfer, wallet, deferred]
        amount:
          type: number
        notes:
          type: string
      required:
        - planId
        - paymentType
...
    Renewal:
      type: object
      properties:
        id:
          type: integer
        subscriberId:
          type: integer
        subscriberNameAr:
          type: string
        subscriberNameEn:
          type: string
        planId:
          type: integer
        planNameAr:
          type: string
        planNameEn:
          type: string
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        paymentType:
          type: string
          enum: [cash, vodafone_cash, bank_transfer, wallet, deferred]
        renewalDate:
          type: string
          format: date-time
        newExpiryDate:
          type: string
          format: date
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - subscriberId
        - planId
        - amount
        - paymentType
        - renewalDate
        - newExpiryDate
    RenewalListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Renewal"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
...